+++ CHANGELOG.md
@@ -51,6 +51,7 @@
   `pcs resource create` ([rhbz#1382597])
 - Fixed validation of options of resource operations in `pcs resource create`
   ([rhbz#1390071])
+- Fixed silent omission of duplicate options ([rhbz#1390066])
 
 [ghissue#124]: https://github.com/ClusterLabs/pcs/issues/124
 [ghissue#125]: https://github.com/ClusterLabs/pcs/issues/125
@@ -67,6 +68,7 @@
 [rhbz#1389443]: https://bugzilla.redhat.com/show_bug.cgi?id=1389443
 [rhbz#1389501]: https://bugzilla.redhat.com/show_bug.cgi?id=1389501
 [rhbz#1389941]: https://bugzilla.redhat.com/show_bug.cgi?id=1389941
+[rhbz#1390066]: https://bugzilla.redhat.com/show_bug.cgi?id=1390066
 [rhbz#1390071]: https://bugzilla.redhat.com/show_bug.cgi?id=1390071
 [rhbz#1394273]: https://bugzilla.redhat.com/show_bug.cgi?id=1394273
 [rhbz#1394846]: https://bugzilla.redhat.com/show_bug.cgi?id=1394846